Renamed Cookbooks to Extensions
This change is complete issofar as symbols in the code are concerned;
class names, variable names, etc, referring to Cookbooks have been
completely extricated and replaced with their Extension equivalents.
This change also includes some places where Cookbooks have been changed
to Extensions in user-facing text, HTML class names, etc, but such
changes are incidental and not to be considered complete with this
commit. As a result, there are likely places in the stylesheets,
JavaScripts, and text which still need updating.